Fixed an overflow in Pipeless Steam Dynamo, fixed Pipeless Steam Hatch Voiding energy when placing or breaking

+++ src/main/java/com/science/gtnl/common/machine/hatch/WirelessSteamDynamoHatch.java
@@ -286,7 +286,7 @@ public class WirelessSteamDynamoHatch extends MTEHatchOutput implements IFluidSt
             }
 
             if (matchedSteamType != null) {
-                int convertedAmount = rawAmount * matchedSteamType.efficiencyFactor;
+                long convertedAmount = (long) rawAmount * (long) matchedSteamType.efficiencyFactor;
 
                 if (!SteamWirelessNetworkManager.addSteamToGlobalSteamMap(ownerUUID, convertedAmount)) {
                     return;
